requirements.txt
torch
torchvision
gradio
seaborn
scikit-learn
tqdm
Pillow
matplotlib
numpy
ucimlrepo

